requirements.txt
torch==2.6.0
torchvision==0.21.0
accelerate==1.4.0
beautifulsoup4==4.10.0
deepspeed==0.16.4
dotenv==0.9.9
einops==0.8.1
flash_attn==2.7.4.post1
huggingface-hub==0.30.2
icecream==2.1.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
jax==0.5.1
Jinja2==3.1.0
ninja==1.11.1.4
numpy==1.21.5
openai==1.76.0
pandas==1.3.5
Pillow==9.0.1
qwen-agent==0.0.20
qwen-vl-utils==0.0.11
tokenizers==0.21.1
transformers==4.50.0